refactor(sayer): sayer.go -> package sayer
